[전반적인 헤더 정리] - inc/drv, inc/gui, inc/mod 에 대한 전반적인 헤더 정리 - <config.h> 에 의존해 GUI 활성화에 대해 헤더들이 영향을 받던 부분을 제거 - 이전에 이 사항에 대해서 적용하는게 좋다고 판단 했다가 다시 안하는게 좋겠다는 판단이 들었음 - 일부 사용하지 않는 파일 삭제

+++ src/gui/painter/yss_painter_rgb565.cpp
@@ -343,7 +343,7 @@ void drawArea(Rgb565 &des, Position areaPos, Size areaSize, Rgb565 &src, Positio
 		(uint16_t)desOffset,			//uint16_t destinationOffset;
 		colorMode::RGB565,	//uint8_t destinationColorMode;
 
-		Size{width, height}	//Size size;
+		Size{(uint16_t)width, (uint16_t)height}	//Size size;
 	};
 	
 	dma2d.lock();
